+++ i2c/src/asynch.rs
@@ -1,49 +1,12 @@
-#![no_std]
-#![cfg_attr(
-    feature = "async",
-    allow(incomplete_features),
-    feature(async_fn_in_trait, impl_trait_projections)
-)]
+use display_interface::{AsyncWriteOnlyDataCommand, DataFormat, DisplayError};
 
-//! Generic I2C interface for display drivers
+use crate::I2CInterface;
 
-#[cfg(feature = "async")]
-pub mod asynch;
-
-use display_interface::{DataFormat, DisplayError, WriteOnlyDataCommand};
-
-/// I2C communication interface
-pub struct I2CInterface<I2C> {
-    i2c: I2C,
-    addr: u8,
-    data_byte: u8,
-}
-
-impl<I2C> I2CInterface<I2C>
+impl<I2C> AsyncWriteOnlyDataCommand for I2CInterface<I2C>
 where
-    I2C: embedded_hal::i2c::I2c,
+    I2C: embedded_hal_async::i2c::I2c,
 {
-    /// Create new I2C interface for communication with a display driver
-    pub fn new(i2c: I2C, addr: u8, data_byte: u8) -> Self {
-        Self {
-            i2c,
-            addr,
-            data_byte,
-        }
-    }
-
-    /// Consume the display interface and return
-    /// the underlying peripherial driver
-    pub fn release(self) -> I2C {
-        self.i2c
-    }
-}
-
-impl<I2C> WriteOnlyDataCommand for I2CInterface<I2C>
-where
-    I2C: embedded_hal::i2c::I2c,
-{
-    fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result<(), DisplayError> {
+    async fn send_commands(&mut self, cmds: DataFormat<'_>) -> Result<(), DisplayError> {
         // Copy over given commands to new aray to prefix with command identifier
         match cmds {
             DataFormat::U8(slice) => {
@@ -52,13 +15,14 @@ where
 
                 self.i2c
                     .write(self.addr, &writebuf[..=slice.len()])
+                    .await
                     .map_err(|_| DisplayError::BusWriteError)
             }
             _ => Err(DisplayError::DataFormatNotImplemented),
         }
     }
 
-    fn send_data(&mut self, buf: DataFormat<'_>) -> Result<(), DisplayError> {
+    async fn send_data(&mut self, buf: DataFormat<'_>) -> Result<(), DisplayError> {
         match buf {
             DataFormat::U8(slice) => {
                 // No-op if the data buffer is empty
@@ -71,17 +35,19 @@ where
                 // Data mode
                 writebuf[0] = self.data_byte;
 
-                slice
-                    .chunks(16)
-                    .try_for_each(|c| {
-                        let chunk_len = c.len();
+                for chunk in slice.chunks(16) {
+                    let chunk_len = chunk.len();
 
-                        // Copy over all data from buffer, leaving the data command byte intact
-                        writebuf[1..=chunk_len].copy_from_slice(c);
+                    // Copy over all data from buffer, leaving the data command byte intact
+                    writebuf[1..=chunk_len].copy_from_slice(chunk);
 
-                        self.i2c.write(self.addr, &writebuf[0..=chunk_len])
-                    })
-                    .map_err(|_| DisplayError::BusWriteError)
+                    self.i2c
+                        .write(self.addr, &writebuf[0..=chunk_len])
+                        .await
+                        .map_err(|_| DisplayError::BusWriteError)?;
+                }
+
+                Ok(())
             }
             DataFormat::U8Iter(iter) => {
                 let mut writebuf = [0; 17];
@@ -98,6 +64,7 @@ where
                     if i == len {
                         self.i2c
                             .write(self.addr, &writebuf[0..=len])
+                            .await
                             .map_err(|_| DisplayError::BusWriteError)?;
                         i = 1;
                     }
@@ -106,6 +73,7 @@ where
                 if i > 1 {
                     self.i2c
                         .write(self.addr, &writebuf[0..=i])
+                        .await
                         .map_err(|_| DisplayError::BusWriteError)?;
                 }